Refactor the User map
